Updated openstack/openstack Project: openstack/tripleo-heat-templates  2c156c1c1370d409e0b58858d78c3c6e26cda6cd
puppet: correct cinder enabled_backends format

This resolves a formatting issue with the Cinder enabled_backends
config file setting. Previously we would potentially construct
an array with an undef value at the end if iscsi was enabled
but ceph was not (this is the case for our current CI job).
When an array formatted like ['tripleo_iscsi', undef] is then
passed to join() in puppet-cinder to construct a string it leaves
us with an extra ',' on the end of the string. This causes
problems in that cinder-volume loads an extra (system default)
cinder volume process which is not expected.

Because Fedora uses LIO as a default it was causing about half
of our CI runs to fail if the tgtadm cinder-volume process
wasn't being chosen by the scheduler.

Closes-bug: #1437708

Change-Id: I3383012cb43792f334fdf789dc13147a3cb5ad63