Updated openstack/openstack Project: openstack/cinder  4d078545bf5c6499634519db961aa5403645fe58
Use cached values for stats on query failures for vmem drivers

This change allows vmem drivers to return previously cached
values for free_capacity_gb and total_capacity_gb if the queries
to the backend for those stats succeed but do not contain
the data.  This can happen when the mgmt plane is under heavy load.

Change-Id: I218a54897361d739c04963b4c36d93620be01fb3
Closes-Bug: 1433990