Updated openstack/openstack Project: openstack/heat-templates  600924e5a3f6eb8e13ac228cf9043d1ed44bf937
Fix hiera deployment JSON generation

We were hitting strange intermittent errors in TripleO where the hiera
JSON file generated by Heat was not even valid JSON, ending with bits of
text which seem out of place, and of varying length, e.g.

"CREATE"}"}

or

"CREATE"}EATE"}

The cause could be that the generation of the file runs multiple times,
and if the first generated JSON is longer than the subsequent one, the
ending of the previous JSON is still present in the file, because the
generation opens the file with O_CREAT (create if not exists) and
O_WRONLY (write only) flags, but not with O_TRUNC (truncate file before
writing). This patch adds the truncation to fix the issue.

Change-Id: Icf184f973decec2eb7de1dee9959b60774d83eb6
Closes-Bug: #1434187