Updated openstack/openstack Project: openstack/manila  6e4f3b31dbee2f08c847436f4b980d3ed8838404
Fix setting of share name and description with manage API

API 'manage' allows us to set name and description for newly managed share.
But it always sets empty strings for both. Root cause for it lays in different
naming of mentioned attrs in API and DB layers.
So, make API 'manage' understand both names (API-like and DB-like) for each of
mentioned things.

Change-Id: I5a752c7960492070c3f5ad5974156c4dea324ac1
Closes-Bug: #1434056