Updated openstack/openstack Project: openstack/nova  9b74fca9244a3dc641cad64e74e3df9e4d471243
Expand valid server group name character set

Names requested for a server group should use the same validation
checking as names in the other APIS. Right now it will not allow
any non alphanumeric characters. Change this to allow all printable
and horizontal whitespace unicode characters. This is the same check
that is done for flavor names, as well as names in v2.1 APIs.
Closes-Bug: #1434335

Change-Id: I6c675fb6c2f40e83499a41ecdccec58d99501433