Updated openstack/openstack Project: openstack/zaqar  3709cdc3caf24662544ae797b91e47f06efc4af3
Allow for internal pipelines to be defined

This patch adds support for defining built-in pipelines that won't
interact with the third-party ones. These pipelines will be always
installed and loaded in the pipeline chain.

Note that, in this patch, the pipelines are added at the beginning of
the stages. In some cases, it may be needed that these pipelines are
executed *after* the store finishes its job.

Partially-implements blueprint: notifications

Change-Id: Ie6a6677cb7ffde37022f26b2231f354e946fb3cf