Updated openstack/openstack Project: openstack/ceilometer  6bd37b0d7efb03dd1f697020b8acf6c5d26233b0
Stop using PYTHONHASHSEED=0 in ceilometer tests

Otherwise we have tests that rely on ordering (for example the one
fixed in this patchset).

Note that this moves the hash seed setting to only be used in docs
and venv tox targets. For some reason there is an issue in the tooling
used by the docs build that tickles a bug in those tools. Tools which
are not part of Ceilometer. To avoid tickling the problem and to move
ceilometer along on this bug, we're just going to mask that.

The venv target is used by the jenkins job that builds docs. It runs
the same build_sphinx setup.py target that the docs target does.

Change-Id: If93ec7934ad59515b821f056ca534d3676ee7e03
Closes-Bug: #1348818