Updated openstack/openstack Project: openstack/cinder  785cc50b973a3d1d3ec48bb13c13c07027731ec1
SMBFS: Fix missing volume provider location

In case of cloning volumes or creating volumes from snapshots, the
volume provider location is not set, making the volumes unusable.

This also leads to disk image leaks, because the delete volume
operation will pass without deleting the disk image if the provider
location is not present.

The SMB driver wraps base RemoteFS methods in order to set locks.
The issue is that it does not return the wrapped method return
value, which signals volume updates, in this case the provider
location.

Change-Id: Id7435676052885d301a48fa5ac224071e18ecf5b
Closes-Bug: #1437271