Updated openstack/openstack Project: openstack/tempest  584ea2de116cda5f956ffaa7bffb88ab5ea325cf
Only use accounts.yaml with locking provider

This commit removes support for using an accounts.yaml file with the
non-locking provider. There is no reason the non-locking provider
would ever need to be used with an accounts.yaml file because the
locking provider supports running serially just fine. (which would be
the sole use case for accounts file with the non-locking provider)
Doing this relegates the non-locking provider to just handle the
legacy config options for providing credentials.

Removing this code path provides the advantage of removing a double
maintenance burden for the non-locking provider, in addition to
simplifying the config and user story around the cred providers. It
also turns out this config permutation was never properly tested and
didn't actually work.

Change-Id: I8088d75897589203264ae29326fe9901c3457cc3