Updated openstack/openstack Project: openstack/python-keystoneclient  389c3ee836192f091bda383b3dcea785d2bfdc68
Allow requesting an unscoped Token

The keystone server understands that specifying unscoped in the scope
section of an auth request means that it should ignore the
default_project_id of a user and return an unscoped token.

This is the client side change to allow requesting these tokens via an
auth plugin.

Change-Id: Iba5ebcea0bf0d8e5a31d552977276fc03e536c67
Implements: bp explicit-unscoped