Updated openstack/openstack Project: openstack/neutron  7b94606f0eb710484153193a7bba37c643e00cee
Set floating IP port status to "N/A"

The status of the port associated with a floating IP
would always show as DOWN. This caused confusion to
operators that weren't aware that this is expected behavior
since the port is only used for an IP allocation.

This commit sets the port status to "N/A" to reflect the fact
that the port associated with a floating IP has no operational
status.

DocImpact
APIImpact

Closes-Bug: #1196851
Change-Id: I2f94afa001b213d61f0e5892aae2e6e6de98fe4c