Updated openstack/openstack Project: openstack/python-ceilometerclient  44af2d4087815d5d451305b549f91ab6130528c0
Add timeout for keystoneclient session

Currently, we only apply timeout on ceilometer api, but not on keystone
api as well, this is not consistent. Keystone api may be slow, or worse,
stuck, then ceilometerclient will not return timely. This issue has caused
availability problem in our environment when ceilometer alarm service
has been running for a long time.

Change-Id: I0d5f82ff9cf4132a1de2f0b649908483326e116e
Closes-Bug: #1436249