Updated openstack/openstack Project: openstack-infra/project-config  d49e76970614a3b7f238700e25b0ec15eead07a7
Move aiopcpu to nova check queue

As the first step in seeing how stable tempest-dsvm-aiopcpu-full is,
make it non-voting on nova's check queue. tempest-dsvm-aiopcpu-full is
currently working (with live migration enabled!) but we don't have
enough data to understand what the success rate is for this job and if
its low enough to become the standard configuration.

The success rate of this job can be monitored via graphite and
logstash.openstack.org

Change-Id: I2152389f31e7ecd5290de77d45df8ffcbdacdb46