Updated openstack/openstack Project: openstack/designate  a88a0abb325ab2049075c7d8a052ead4bd91f685
Fix ownership issue with Zone's created on set fip

When a user sets a FloatingIP a Reverse Zone is created that's owned by the
system but if no managed_resource_tenant_id option was set then the Zone would
get owned by the context.tenant_id when create_domain() is called in
set_floatingip.

We fix it by changing the default managed_resource_tenant_id to
"00000000-0000-0000-0000-000000000000".

Closes-Bug: #1437060

Change-Id: I491309d92a9e3fba2f42af1ae85372b6258d53f3