Updated openstack/openstack Project: openstack/tuskar  0d4d98d6d5e0e97270d32206da2525dc97c560c2
Add registry_path to the Role object

Now that we have decoupled the Role name from the name of the file that
defines it, there is no longer a link between a Role and the resource
registry type that it defines. For instance, there is no link between
'Compute' and the relevant entry in the resource registry:

OS::TripleO::Compute: puppet/compute-puppet.yaml

When loading the role, capture its file as part of the Role object so
that we can dereference the Role name against the resource registry,
regardless of name.

Change-Id: I0d36355dbf4e296584d05bf5a137926f85a78b43