Updated openstack/openstack Project: openstack/tuskar  451f8433497249225753bde8de16fe912ed898ac
Include missing aliases in resource registry

When loading the seed, resource registry entries referencing other files
was recently added to the final environment. However, other aliases were
skipped and this should not be the case.

The test_load_roles is the primary verification, but having another
example in the test_parser is helpful for awareness.

Change-Id: I28c8175140f45a1769eb372edcffb3f78bdd780f