Updated openstack/openstack Project: openstack/ceilometer  8775138bb97665eab28e85c621420ef9e1c07c74
Avoid a error when py27 and py-mysql tests run in sequence

An imcompletely MagicMock can lead to:

  File "ceilometer/network/statistics/opencontrail/client.py", line
  107, in _log_res
      LOG.debug(''.join(dump))
      TypeError: sequence item 3: expected string, MagicMock found

When the py-mysql tox test target is run after the py27 target.

This happens because the Mock is not sufficiently formed for the
places where it gets used. The "dump" above needs to have a value
for 'reason'.

Change-Id: I36fd406a06958e795ecf9aa9ff933eddaa979fc8
Closes-Bug: #1438300