Updated openstack/openstack Project: openstack/nova  3a7cfdfca6995e62b256be5b46da95d10c3950d3
Vmware:Find a SCSI adapter type for attaching iSCSI disk

It needs to find out a proper SCSI adapter before attaching an
iSCSI disk. Current logic is to get an existing disk and use its
adapter type to create and attach iSCSI disk. So an SCSI disk
would be created under an IDE controller if the existing
disk is an IDE disk. This fix will find out a SCSI controller from
the vm and return its corresponding adapter type.

Change-Id: I33b6f3a88a5253755971622b4f05a30f48003331
Closes-bug: #1386511