Updated openstack/openstack Project: openstack/nova  620963b2d3361e72121417c0fd4e9169f6884460
Fix copy configdrive during live-migration on HyperV

This bug was introduced due to an oversight when fixing
https://launchpad.net/bugs/1322096 . In that fix, we use the
"config_drive_cdrom" value from the conf to determine if a
there is a iso configdrive attached to the instance. This
approach is wrong due to the fact that that config value can be
modified by the user. If that happens, instances that still have
iso configdrives will be ommitted.

This path fixes this problem by checking the instance itself if
it has any iso attached to it.

Closes-Bug: #1427467
Change-Id: Ia64389b4405109d7db6fbf9c000a0136e9862298