Updated openstack/openstack Project: openstack/cinder  621cde6664fe1b8176d119b34eb057a63188c842
Enhance VNX Cinder volume creation logic

Sometimes sysadmin needs to remove broken disks and replace them
with some new disks. Before the disks are replaced and fully rebuilt,
the pool would be in "Faulted" state, and all LUNs created in this
pool would be in "Faulted" state as well.
A LUN in this "Faulted" state is still OK for IO access.

Current implementation of create_volume in VNX Cinder Driver will
wait until the VNX LUN gets into Ready State or timeout.
Thus the new volume in the back end corresponding to the pool will
be stuck at creating state for a long time before disks are replaced.
This reduces the high availability of the back end.

The change allows new volume to get into available state as long as
the corresponding VNX LUN can serve IO.

Change-Id: Ifb152b3b9c59ce0d5c875129a7bf0503740d5a3a
Closes-Bug: 1436241