Updated openstack/openstack Project: openstack/cinder  8e6c4cbb891264cac5f6415c7d79bc0be72f29cf
Fix HNAS iSCSI driver error on LUN creation

If chap_enabled is 'False', in the first initialization of the driver (before
the target is created on HNAS) the driver fails to create volumes. This patch
fix that by returning the right value when 'get_targetiqn' is called in this
situation.

Change-Id: I899717c0e7e4ff800cecc8ffa6e3a81a47e96e9f
Closes-Bug: #1430782