Updated openstack/openstack Project: openstack/horizon  302005a42662308c154c4519f769ba1ff3470b7c
Modified hzSelectAll to select all visible checkboxes in table

The hzSelectAll directive should be used with the 'st-safe-src'
attribute and should be passed in the displayed collection. When
used with filtering and pagination, clicking on the select all
checkbox will only mark the checkboxes of visible rows in the
table.

Change-Id: I1e2971f631bf11f5bcff85754b4927212f871f5e
Closes-Bug: #1438351