Updated openstack/openstack Project: openstack/neutron-fwaas  3a6cca8dd5173f69a16d956f500582ddc1f54a31
FWaaS: Refactor callback mechanism

As part of the cleanup work for L3 agent, where we are switching to
a different callback mechanism, and cleaning up the *aaS agents, there
are some minor cleanups that can be done for FWaaS.

First, there are no callbacks defined yet for FWaaS, so we can just
remove the old mechanism. That involves eliminating the AdvancedService
base class, and removing the add_observer() method for callback
registration.

Second, the driver doesn't need access to the l3_agent, we can remove
the argument, when constructing the FirewallService.

Change-Id: Id509f10f126931d01c272b14ada4ad34cbf1d858
Partial-Bug: #1433552