Updated openstack/openstack Project: openstack/neutron  629d1d44ee3d223abc51fc78602d2b0cddfc2bca
Fix test case for DHCP agent interface restart

One of the new test cases in the recent DHCP
interface patch[1] was supposed to confirm that
the driver wouldn't be restarted if the IP address
stayed the same. However, it wasn't matching the
device ID of the agent so it was never making it
to that conditional.

This patch just fixes that UT so it's exercising
the right code path.

1. c4a7447e2d659b3a240a62ae9d34e6e0b9cee7a3

Change-Id: I8735c6e533d6b486c32cfded2c22eac8a25c855d