Updated openstack/openstack Project: openstack/manila  b196f9b80385b94f67fd3d6772a4460bb2303914
Add segmentation ID checks for different segmentation types

The segmentation approach used by the network plugin associated with a share
server was not declared to the driver. This made it impossible for a driver
to know whether a network used VLAN, GRE, VXLAN, or any other segmentation
approach, and introduced the possibility of an incorrect segmentation
approach to be applied to a share server.
This patchset makes the network_type field available to the driver, and
checks that segmentation id has valid values for different network_types.
If the segmentation id has an incorrect value, it fails with a clean
error message.

Partial-Bug: 1438825

Change-Id: I9f20a26e9c10d49ca3c82a46f49bdb5c65fd4a24