Updated openstack/openstack Project: openstack/nova  739f4712644a0b14f1be76c2aa7316bcf9bddfc8
Better power_state logging in _sync_instance_power_state

Commit aa1792eb4c1d10e9a192142ce7e20d37871d916a added more verbose
logging of the various database and hypervisor states when
_sync_instance_power_state is called (which can be called from
handle_lifecycle_event - triggered by the libvirt driver, or from the
_sync_power_states periodic task).

This change adds logging of when the power_state in the database does
not match the power_state from the hypervisor, and then the power_state
in the database is updated to match what's in the hypervisor.

Also, we save the original database power_state for logging when we have
cases of stopping an active instance due to vm_state / power_state
conflicts, or stopping a "stopped" instance but the hypervisor says is
still running.

This is all needed because debugging the various things that can hit
this code at different times and cause problems is, well, terrible.

Change-Id: I297ca0037f43535c80afa8a4b5086ccb30558dc3