Updated openstack/openstack Project: openstack/ironic  767ba3b2b90065bd895e29041901a82233b33226
Prevent updating the node's driver if console is enabled

This patch is preventing a node from updating it's driver if the console
is enabled. That's because the new driver won't be able to stop the
console started by the previous one. In the future we might look at
stopping the console instead of just failing to update the node, but
since it requires some plumbing work just stopping the update seems good
to prevent this bug from sneaking in the RC1 release.

Closes-Bug: #1436884
Change-Id: I1f8778142faaac608fd9eed4d0c40d0d372454db