Updated openstack/openstack Project: openstack/oslo.policy  b5f07dfe4cd4a5d12c7fecbc3954694d934de642
Avoid reloading policy files in policy.d for every call

We added policy files in policy.d, currently, the code will
reload the policy in policy.d every enforce been called.

This patch caches the file mtime of the most newest file in
the directory (or the directory itself if it is empty) and uses
that to detect if we need to reload files in the policy.d
directory.

Closes-Bug: 1437992
Co-Authored-By: Eli Qiao <liyong.qiao@intel.com>
Change-Id: Ia8c273c8566a4b472cb8807b1576ef1dee8ef054