Updated openstack/openstack Project: openstack/heat-translator  3d20aa6df28097fc35745744dc74475c3fa457fb
Add scripts to set up monitoring

This set of scripts installs and configures collectd and rsyslog on a server.
They are set up as service to send monitoring data (collectd) and system logs
(rsyslog) to a logstash server, so the config scripts expects the IP for
logstash in the environment variable logstash_ip.
Collectd uses the udp protocol and rsyslog uses the RELP protocol.

Implements: blueprint translate-tosca-monitoring-usecase
Change-Id: Id594111df2af4e63027312229eb60ac167ed57df