Updated openstack/openstack Project: openstack/heat-translator  f7a71d7ebdbbdd64f7869ad7ad92139f89d66e81
Add scripts for logstash, elasticsearch, kibana

This set of scripts sets up the infrastructure for managing the
monitoring data.  Logstash receives performance metric from collectd
via the udp protocol and logs from rsyslog via the RELP protocol.
They should be coming from the nodejs and mongodb servers.
Logstash then forwards the data stream to elasticsearch for storage and
indexing.  The script for logstash configuration expects the IP address
for the elasticsearch server as the environment elasticsearch_ip.
The scripts are designed to install these as services on separate
servers.

Kibana is the web app to display the monitoring data.
It should be accessible at:
http://<kibana_ip>:5601

Implements: blueprint translate-tosca-monitoring-usecase
Change-Id: I117a26a435bd81f496b1fa58c667316d9cc7ffe5