Updated openstack/openstack Project: openstack/cinder  ce21caef14689dc955099503745795432f5fc930
Remove LP bug ref in remove_iscsi_device

Some like referencing bugs in code comments, others don't.  In
general we try and avoid it, but I certainly don't agree with
blocking a patch solely for that reason.  So instead this patch
tries to modify the previous commit to fix it up rather than
having a contest of wills to see who will bend first.

Change-Id: I98bca9e3a98383c994bc49426a71d944e10fae17