Updated openstack/openstack Project: openstack/manila  3e6ef365c9725e6d0a6c77c5d4fccbfd550a7af6
NetApp cDOT drivers should not start without aggregates

A common user error with the single-SVM cDOT driver is failing to
assign aggregates to the storage virtual machine (aka vserver) used
by the driver.  The driver should fail fast if there are no available
aggregates for share provisioning.

We want the multi-SVM driver to fail fast as well, if only for
consistency and to alert an admin of a problem with the
netapp_aggregate_name_search_pattern regex in manila.conf.

Change-Id: I2a2015bd07b68177b0a5e7ac322ffa50b3b41c7d
Closes-Bug: #1433871