Updated openstack/openstack Project: openstack/cinder  bef2256e931cfa0756de874ac18fdfc12e32b7fc
Fix the format of the system name in VMAX driver

There is a change in the format for system name in SMI-S for VMAX V3.
It was "SYMMETRIX+xxxxxx" previously and now it is changed to
"SYMMETRIX-+-xxxxxx". In this patch, the VMAX driver was modified
accordingly to support this format.

Change-Id: Ie7e349db13d6ad2562a296d3d0cb66089a54b800
Closes-Bug: #1437490