Updated openstack/openstack Project: openstack/heat  6e8b31fae16774b7b929a4b3a981aefed903058d
Add separate requirements.txt to integration tests

and use it during tox run. This should speed up the tox venv creation
somewhat, as the new requirements is shorter, thus less packages are
installed or C extensions compiled.

Also this is the first step to ease packaging of the integration tests
as a separate tool.

Change-Id: I914eb675e7f025e03a22719d5dff8ed741fbd37d