Updated openstack/openstack Project: openstack-infra/shade  d2d5acae2545bb3daef8a86a30d3cacd5c0c9e8d
Add 'rebuild' to shade

Adding the ability to rebuild instances via shade.  While this is not
used by nodepool currently, there is some interest in evaluating whether
or not it could be a viable replacement for the current delete /
create strategy employed by nodepool.  Having rebuild in shade now will
make benchmarks a little simpler.

Change-Id: I12a00b71d248ce77e038dce1642d229587487886