Updated openstack/openstack Project: openstack-infra/shade  23c376431d939e01a55ff081936e8606a3e7867c
Fix a use of in where it should be equality

The pattern for most other comparisons against name_or_id are if
name_or_id in (name, id) - but in this case, because we're doing the
name exclusion parameter as well, we split them. However, we kept the
in, even though it should be testing for equality. This can cause some
failures that make people sad.

Change-Id: If33bfe5c1e32b36c27c32cfcba4c07dcf01a4e74