Updated openstack/openstack Project: openstack/horizon  e561cef4fb34c08367e0b0b3c9d7074a37edb8c8
Use WEBROOT setting for the 500 page home button

The home button on the 500 page currently links to "/" which is not
always the true app root, in settings we define a WEBROOT
variable so that we can inform the app of its actual location.

This patch uses that variable to set href location for the home link
on the 500 page, so it will be acurate for all cases.

To make this possible the load_conf template tag now returns conf
itself instead of conf.HORIZON_CONFIG giving the template access to
more of the settings including WEBROOT.

Change-Id: I9d42ac8eeb7f2bf558ba86440fd5ae85ef591568
Closes-Bug: 1427903