Updated openstack/openstack Project: openstack/glance_store  a5056336ec5acc91f34c2c9a3e155472052f1b40
gridfs: add pymongo to test-requirements and update tests

pymongo is an optional requirement for glance_store if you're not using
the gridfs backend, but for tests we should have it in
test-requirements.txt with other optional backend dependencies like boto
for s3 and oslo.vmware for the vmware backend.

This adds pymongo to test-requirements.txt and removes the conditional
imports and skips in the gridfs tests.

Change-Id: I9332fc5b27e35ed6194eedf5ccd1586c3ba9096e