Updated openstack/openstack Project: openstack/cinder  ebe3234d9efc949979dd8097ab25e70b6c5ff09b
Be safe with getting attachment

The volume manager wasn't being safe with trying to pull an
attachment out of the attachments list.  It's possible someone
is calling the manager to detach a volume that has no attachments.

This patch raises an InvalidVolume exception if we detect that
there are no attachments left to detach.

Change-Id: I684ed870d13a4efc6f64b0144c987e355428bc54
Closes-Bug: #1435574